This adds a date picker for datetime values too
I havent been able to convert from the number representation of datetimes to the Date + Time representation.
I think that it should be done in the "select-item" component, and I should probably use the "datetime-value" function,
but that function needs a repeated-task? parameter. So maybe I should do it via "property-value-content" or "inline-text-cp"
which are both called by "select-item"